backend/python_ai_services/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
google-generativeai==0.8.3
python-dotenv==1.0.1
requests==2.32.3
websockets==14.1
